Fix failing tests

+++ src/gpu/sentry_gpu_common.c
@@ -66,19 +66,19 @@ sentry__get_gpu_context(void)
 
     if (gpu_info->vendor_id != 0) {
         sentry_value_set_by_key(gpu_context, "vendor_id",
-            sentry_value_new_int32(gpu_info->vendor_id));
+            sentry_value_new_int32((int32_t)gpu_info->vendor_id));
     }
 
     // Add device ID
     if (gpu_info->device_id != 0) {
         sentry_value_set_by_key(gpu_context, "device_id",
-            sentry_value_new_int32(gpu_info->device_id));
+            sentry_value_new_int32((int32_t)gpu_info->device_id));
     }
 
     // Add memory size
     if (gpu_info->memory_size > 0) {
         sentry_value_set_by_key(gpu_context, "memory_size",
-            sentry_value_new_int64(gpu_info->memory_size));
+            sentry_value_new_int64((int64_t)gpu_info->memory_size));
     }
 
     // Add driver version
